# Refactor code to reduce minified size