Closes #16.

Summary of changes:

* Configure uglifyjs to mangle private properties, which are denoted with a leading `_` character.
* Pull out token string literals are vars to avoid duplication
* Modify lexing identifiers and numbers to use nested conditionals instead of lookup tables.

#### Results

Size in bytes:

* before: 22179
* **after: 18128**
* **reduction: 18.3%**

Performance is surprisingly better in every case:

before:

```
Mean time: 0.000810msec Parser#single_expr x 1,235,272 ops/sec ±0.35% (91 runs sampled)
Mean time: 0.001408msec Parser#single_subexpr x 710,339 ops/sec ±0.91% (92 runs sampled)
Mean time: 0.021024msec Parser#deeply_nested_50 x 47,564 ops/sec ±0.45% (94 runs sampled)
Mean time: 0.046507msec Parser#deeply_nested_50_index x 21,502 ops/sec ±0.44% (94 runs sampled)
Mean time: 0.002269msec Parser#basic_list_projection x 440,765 ops/sec ±0.47% (93 runs sampled)
```

after:

```
Mean time: 0.000149msec Parser#single_expr x 6,692,389 ops/sec ±0.41% (94 runs sampled)
Mean time: 0.000304msec Parser#single_subexpr x 3,284,377 ops/sec ±0.89% (90 runs sampled)
Mean time: 0.009051msec Parser#deeply_nested_50 x 110,479 ops/sec ±0.54% (94 runs sampled)
Mean time: 0.027004msec Parser#deeply_nested_50_index x 37,031 ops/sec ±0.61% (91 runs sampled)
Mean time: 0.000677msec Parser#basic_list_projection x 1,476,137 ops/sec ±0.60% (92 runs sampled)
```

Note in the first benchmark that's 5.4 times faster with these changes.